Placed OS functions for driver in a seperate cpp file
